+++ java/client/src/main/java/glide/api/commands/PubSubClusterCommands.java
@@ -3,7 +3,6 @@ package glide.api.commands;
 
 import glide.api.models.GlideString;
 import java.util.concurrent.CompletableFuture;
-import lombok.NonNull;
 
 /**
  * Supports commands for the "Pub/Sub" group for a cluster client.
@@ -51,11 +50,8 @@ public interface PubSubClusterCommands {
     /**
      * Lists the currently active shard channels.
      *
-     * @apiNote When in cluster mode, the command is routed to all nodes, and aggregates the response
-     *     into a single array.
      * @see <a href="https://valkey.io/commands/pubsub-shardchannels/">valkey.io</a> for details.
-     * @return An <code>array</code> of currently active shard channels matching the given pattern. If
-     *     no pattern is specified, all active shard channels are returned.
+     * @return An <code>array</code> of all active shard channels.
      * @example
      *     <pre>{@code
      * String[] result = client.pubsubShardChannels().get();
@@ -67,11 +63,8 @@ public interface PubSubClusterCommands {
     /**
      * Lists the currently active shard channels.
      *
-     * @apiNote When in cluster mode, the command is routed to all nodes, and aggregates the response
-     *     into a single array.
      * @see <a href="https://valkey.io/commands/pubsub-shardchannels/">valkey.io</a> for details.
-     * @return An <code>array</code> of currently active shard channels matching the given pattern. If
-     *     no pattern is specified, all active shard channels are returned.
+     * @return An <code>array</code> of all active shard channels.
      * @example
      *     <pre>{@code
      * GlideString[] result = client.pubsubShardChannelsBinary().get();
@@ -83,34 +76,28 @@ public interface PubSubClusterCommands {
     /**
      * Lists the currently active shard channels.
      *
-     * @apiNote When in cluster mode, the command is routed to all nodes, and aggregates the response
-     *     into a single array.
      * @see <a href="https://valkey.io/commands/pubsub-shardchannels/">valkey.io</a> for details.
      * @param pattern A glob-style pattern to match active shard channels.
-     * @return An <code>array</code> of currently active shard channels matching the given pattern. If
-     *     no pattern is specified, all active shard channels are returned.
+     * @return An <code>array</code> of currently active shard channels matching the given pattern.
      * @example
      *     <pre>{@code
      * String[] result = client.pubsubShardChannels("channel*").get();
      * assert Arrays.equals(result, new String[] { "channel1", "channel2" });
      * }</pre>
      */
-    CompletableFuture<String[]> pubsubShardChannels(@NonNull String pattern);
+    CompletableFuture<String[]> pubsubShardChannels(String pattern);
 
     /**
      * Lists the currently active shard channels.
      *
-     * @apiNote When in cluster mode, the command is routed to all nodes, and aggregates the response
-     *     into a single array.
      * @see <a href="https://valkey.io/commands/pubsub-shardchannels/">valkey.io</a> for details.
      * @param pattern A glob-style pattern to match active shard channels.
-     * @return An <code>array</code> of currently active shard channels matching the given pattern. If
-     *     no pattern is specified, all active shard channels are returned.
+     * @return An <code>array</code> of currently active shard channels matching the given pattern.
      * @example
      *     <pre>{@code
      * GlideString[] result = client.pubsubShardChannels(gs.("channel*")).get();
      * assert Arrays.equals(result, new GlideString[] { gs("channel1"), gs("channel2") });
      * }</pre>
      */
-    CompletableFuture<GlideString[]> pubsubShardChannels(@NonNull GlideString pattern);
+    CompletableFuture<GlideString[]> pubsubShardChannels(GlideString pattern);
 }